Move backup unload logic to an event
Fixes #15260
