+++ src/vs/platform/backup/electron-main/backupMainService.ts
@@ -10,6 +10,8 @@ import Uri from 'vs/base/common/uri';
 import { readdirSync } from 'vs/base/node/extfs';
 import { IBackupWorkspacesFormat, IBackupMainService } from 'vs/platform/backup/common/backup';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
+import { ILifecycleService } from 'vs/code/electron-main/lifecycle';
+import { VSCodeWindow } from 'vs/code/electron-main/window';
 
 export class BackupMainService implements IBackupMainService {
 
@@ -21,11 +23,27 @@ export class BackupMainService implements IBackupMainService {
 	private workspacesJsonContent: IBackupWorkspacesFormat;
 
 	constructor(
-		@IEnvironmentService environmentService: IEnvironmentService
+		@IEnvironmentService environmentService: IEnvironmentService,
+		@ILifecycleService lifecycleService: ILifecycleService
 	) {
 		this.backupHome = environmentService.backupHome;
 		this.workspacesJsonPath = environmentService.backupWorkspacesPath;
+
+		if (lifecycleService) {
+			lifecycleService.onAfterUnload(this.onAfterUnloadWindow.bind(this));
+		}
+
 		this.loadSync();
+	}
+
+	private onAfterUnloadWindow(vscodeWindow: VSCodeWindow) {
+		if (vscodeWindow.openedWorkspacePath) {
+			// Clear out workspace from workspaces.json if it doesn't have any backups
+			const workspaceResource = Uri.file(vscodeWindow.openedWorkspacePath);
+			if (!this.hasWorkspaceBackup(workspaceResource)) {
+				this.removeWorkspaceBackupPathSync(workspaceResource);
+			}
+		}
 	}
 
 	public getWorkspaceBackupPaths(): string[] {
